requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.1,<3.0.0
pydantic-settings>=2.1.0
crewai==0.28.8
crewai-tools==0.1.6
python-pptx==0.6.23
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
Pillow==10.1.0
requests==2.31.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.2
google-cloud-error-reporting==1.12.0

# Google AI and Drive APIs
google-generativeai>=0.8.3
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Document Generation
python-docx>=1.1.2

# RAG and Vector Database Dependencies
chromadb>=0.4.22,<0.5.0
# CrewAI 0.28.8 requires langchain >=0.1.10,<0.2.0
langchain>=0.1.10,<0.2.0
langchain-community==0.0.34
langchain-chroma==0.1.0
sentence-transformers>=2.2.2
faiss-cpu==1.7.4
tiktoken>=0.7.0,<0.8.0
numpy==1.24.3

# LangGraph Workflow Orchestration
langgraph==0.0.20

# Advanced Document Processing
pypdf2==3.0.1
python-magic==0.4.27
spacy==3.7.2
nltk==3.8.1
chardet==5.2.0

# Machine Learning and Analytics
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Production Monitoring and Performance
psutil==5.9.6
prometheus-client==0.19.0
redis==5.0.1
celery==5.3.4

# Async and Concurrency
aiofiles==23.2.1
asyncio-mqtt==0.16.1
websockets==12.0

# Data Processing and Analysis (deduplicated)

# Security and Validation
cryptography>=41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing and Quality Assurance
pytest==8.3.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
